separating configs into files
