+++ blockchain/tokens/goerli.ts
@@ -1,0 +1,3 @@
+export const ACCOUNT_GUARD_FACTORY_GENESIS_GOERLI = 8420373
+export const AAVE_V3_POOL_GENESIS_GOERLI = 8294332
+export const AAVE_V2_LENDING_POOL_GENESIS_GOERLI = 7480475
